requirements.txt
python-dotenv 
groq 
swarms